fix(cubesql): Use pushdown-pullup scheme for FilterSimplifyReplacer (#9278)
This should avoid unexpected unifications of simplified expression in unrelated context. For example, same expression can be present in filter and projection, but due to over-unification in filters it can receive different representation in projection, and break aliases later, during extraction

Supporting changes:

* Remove unnecessary Vec+Box for rules

* Add NOW function template
A couple of unit tests with NOW and wrapper were relying on over-unification, and broke without a function template.

* Bump default rewrite iteration limit to 500
With new rules `select_many_filters` benchmark was hitting iteration limit